Created most pages and added a header to each page
I started HTML files for the following pages: home, goods, services, listings, login, and profile. To each file, I created a very rough header that links to the other pages. There is also a main.js file that if you modify the first line, simulates the difference (for the header only) between the user being logged in and not logged in.