Created sample listings for the home page
I created sample listings for the home page. They are all hard coded at the moment but are in place for when they will be updated. Nothing is set in stone yet so all listings and sections are free to be changed in any way needed. I also shrunk the header size to better fit the screen and the new listings section.