Lab 1 Front End
I created a constitution folder in the repo with the contents for lab 1. There is no back end / JS, but the front end is in place and is ready to be expanded (hopefully) when all of the other information is in place. I created a README file where I put some notes and some things left to do.