Renamed single_auction.html to index.html and created PHP + JS to communicate with the database
I first renamed the single_auction.html to be index.html. This is useful so when redirecting to the new directory we do not need to specify the html file as it defaults to index.html. Second, I created a JS and two PHP files which work on communication with the SQL database. The first gathers information from the bidData and auctionData page and populates the information. The second is ran when the user places a bid and adds it if all criteria is met

Issue:
https://github.com/orgs/RPI-ITWS/projects/54?pane=issue&itemId=80611477